ConflictResolver: Refactoring.
